+++ tests/osu.rs
@@ -3,7 +3,6 @@
 extern crate rosu_pp;
 
 use rosu_pp::Beatmap;
-use std::fs::File;
 
 struct MapResult {
     map_id: u32,
@@ -12,9 +11,27 @@ struct MapResult {
     pp: f32,
 }
 
-#[test]
-fn osu() {
-    let margin = if cfg!(feature = "no_sliders_no_leniency") {
+macro_rules! assert_result {
+    ($kind:expr => $result:expr, $margin:expr, $expected:ident, $map_id:ident, $mods:ident) => {
+        assert!(
+            ($result - $expected).abs() < $margin * $expected,
+            "\n{kind}:\n\
+                Calculated: {calculated} | Expected: {expected}\n \
+                => {margin} margin ({allowed} allowed)\n\
+                [map {map} | mods {mods}]\n",
+            kind = $kind,
+            calculated = $result,
+            expected = $expected,
+            margin = ($result - $expected).abs(),
+            allowed = $margin * $expected,
+            map = $map_id,
+            mods = $mods
+        );
+    };
+}
+
+fn margin() -> f32 {
+    if cfg!(feature = "no_sliders_no_leniency") {
         0.0075
     } else if cfg!(feature = "no_leniency") {
         0.0025
@@ -22,7 +39,13 @@ fn osu() {
         0.001
     } else {
         unreachable!()
-    };
+    }
+}
+
+#[cfg(not(any(feature = "async_std", feature = "async_tokio")))]
+#[test]
+fn osu_sync() {
+    let margin = margin();
 
     let star_margin = margin;
     let pp_margin = margin;
@@ -35,7 +58,7 @@ fn osu() {
             pp,
         } = result;
 
-        let file = match File::open(format!("./maps/{}.osu", map_id)) {
+        let file = match std::fs::File::open(format!("./maps/{}.osu", map_id)) {
             Ok(file) => file,
             Err(why) => panic!("Could not open file {}.osu: {}", map_id, why),
         };
@@ -47,51 +70,53 @@ fn osu() {
 
         let result = rosu_pp::OsuPP::new(&map).mods(*mods).calculate();
 
-        assert!(
-            (result.stars() - stars).abs() < star_margin * stars,
-            "\nStars:\n\
-            Calculated: {calculated} | Expected: {expected}\n \
-            => {margin} margin ({allowed} allowed)\n\
-            [map {map} | mods {mods}]\n",
-            calculated = result.stars(),
-            expected = stars,
-            margin = (result.stars() - stars).abs(),
-            allowed = star_margin * stars,
-            map = map_id,
-            mods = mods
-        );
-
-        assert!(
-            (result.pp() - pp).abs() < pp_margin * pp,
-            "\nPP:\n\
-            Calculated: {calculated} | Expected: {expected}\n \
-            => {margin} margin ({allowed} allowed)\n\
-            [map {map} | mods {mods}]\n",
-            calculated = result.pp(),
-            expected = pp,
-            margin = (result.pp() - pp).abs(),
-            allowed = pp_margin * pp,
-            map = map_id,
-            mods = mods
-        );
+        assert_result!("Stars" => result.stars(), star_margin, stars, map_id, mods);
+        assert_result!("PP" => result.pp(), pp_margin, pp, map_id, mods);
     }
+}
+
+#[cfg(feature = "async_tokio")]
+#[test]
+fn osu_async_tokio() {
+    tokio::runtime::Runtime::new()
+        .expect("could not start runtime")
+        .block_on(async {
+            let margin = margin();
+
+            let star_margin = margin;
+            let pp_margin = margin;
+
+            for result in RESULTS {
+                let MapResult {
+                    map_id,
+                    mods,
+                    stars,
+                    pp,
+                } = result;
+
+                let file = match tokio::fs::File::open(format!("./maps/{}.osu", map_id)).await {
+                    Ok(file) => file,
+                    Err(why) => panic!("Could not open file {}.osu: {}", map_id, why),
+                };
+
+                let map = match Beatmap::parse(file).await {
+                    Ok(map) => map,
+                    Err(why) => panic!("Error while parsing map {}: {}", map_id, why),
+                };
+
+                let result = rosu_pp::OsuPP::new(&map).mods(*mods).calculate();
+
+                assert_result!("Stars" => result.stars(), star_margin, stars, map_id, mods);
+                assert_result!("PP" => result.pp(), pp_margin, pp, map_id, mods);
+            }
+        });
 }
 
 #[cfg(feature = "async_std")]
 #[test]
-fn osu_async() {
-    use async_std::{fs::File, task};
-
-    task::block_on(async {
-        let margin = if cfg!(feature = "no_sliders_no_leniency") {
-            0.0075
-        } else if cfg!(feature = "no_leniency") {
-            0.0025
-        } else if cfg!(feature = "all_included") {
-            0.001
-        } else {
-            unreachable!()
-        };
+fn osu_async_std() {
+    async_std::task::block_on(async {
+        let margin = margin();
 
         let star_margin = margin;
         let pp_margin = margin;
@@ -104,45 +129,20 @@ fn osu_async() {
                 pp,
             } = result;
 
-            let file = match File::open(format!("./maps/{}.osu", map_id)).await {
+            let file = match async_std::fs::File::open(format!("./maps/{}.osu", map_id)).await {
                 Ok(file) => file,
                 Err(why) => panic!("Could not open file {}.osu: {}", map_id, why),
             };
 
-            let map = match Beatmap::parse_async(file).await {
+            let map = match Beatmap::parse(file).await {
                 Ok(map) => map,
                 Err(why) => panic!("Error while parsing map {}: {}", map_id, why),
             };
 
             let result = rosu_pp::OsuPP::new(&map).mods(*mods).calculate();
 
-            assert!(
-                (result.stars() - stars).abs() < star_margin * stars,
-                "\nStars:\n\
-                Calculated: {calculated} | Expected: {expected}\n \
-                => {margin} margin ({allowed} allowed)\n\
-                [map {map} | mods {mods}]\n",
-                calculated = result.stars(),
-                expected = stars,
-                margin = (result.stars() - stars).abs(),
-                allowed = star_margin * stars,
-                map = map_id,
-                mods = mods
-            );
-
-            assert!(
-                (result.pp() - pp).abs() < pp_margin * pp,
-                "\nPP:\n\
-                Calculated: {calculated} | Expected: {expected}\n \
-                => {margin} margin ({allowed} allowed)\n\
-                [map {map} | mods {mods}]\n",
-                calculated = result.pp(),
-                expected = pp,
-                margin = (result.pp() - pp).abs(),
-                allowed = pp_margin * pp,
-                map = map_id,
-                mods = mods
-            );
+            assert_result!("Stars" => result.stars(), star_margin, stars, map_id, mods);
+            assert_result!("PP" => result.pp(), pp_margin, pp, map_id, mods);
         }
     })
 }