+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## Upcoming
 
-# v0.1.1
+- No changes so far
+
+# v0.2.0
+
+- Async beatmap parsing through features `async_tokio` or `async_std`
+- Hide various parsing related types further inwards, i.e. `rosu_pp::parse::some_type` instead of `rosu_pp::some_type`
+  - Affected types: `DifficultyPoint`, `HitObject`, `Pos2`, `TimingPoint`, `HitObjectKind`, `PathType`, `HitSound`
+
+## v0.1.1
 
 - parse:
   - Efficiently handle huge amounts of curvepoints